Add files via upload
These various kml files are in xml format these carry both data and metadata (data about the data) for TRAILS, THE PERIMETER OF OAHU, PARKS/RESERVES POLYGONS
our latest r script relies on them to render the enclosed png file.  We have much work to do! but we have a map drawn with kml data that is ready for cell phone ping data.  

once we have our core functionality then we can work on making plots pretty and then looping on layer to produce a gif or 'video' of cell phone pings moving about the parks....